fix: Predict components route
